chore: Update YC Deploy Container to Instance Group action
This commit updates the YC Deploy Container to Instance Group action. It changes the name, description, and author fields in the action.yml file. Additionally, it updates the required inputs' descriptions to use double quotes instead of single quotes for consistency. The runs field in the action.yml file is also updated to use Node.js version 20 instead of version 16.